Move Loading component into the new utils folder
