+++ web/html/src/manager/virtualization/guests/edit/guests-edit.js
@@ -8,7 +8,7 @@ const React = require('react');
 const _isEqual = require('lodash/isEqual');
 const { TopPanel } = require('components/panels/TopPanel');
 const MessagesUtils = require('components/messages').Utils;
-const { Loading } = require('components/loading/loading');
+const { Loading } = require('components/utils/Loading');
 const { getOrderedItemsFromModel } = require('components/input/FormMultiInput');
 const { GuestProperties } = require('../guest-properties');
 const GuestNicsPanel = require('../properties/guest-nics-panel');
