+++ web/html/src/manager/virtualization/guests/guest-properties.js
@@ -8,7 +8,7 @@ const { Text } = require('components/input/Text');
 const { Select } = require('components/input/Select');
 const Validation = require('components/validation');
 const MessagesUtils = require('components/messages').Utils;
-const { Loading } = require('components/loading/loading');
+const { Loading } = require('components/utils/Loading');
 const { guestNicsPanel } = require('./properties/guest-nics-panel');
 const { guestDisksPanel } = require('./properties/guest-disks-panel');
 const { GuestPropertiesForm } = require('./properties/guest-properties-form');
